Update purchase API files

+++ server/src/main/java/com/express/inventory/api/purchases/PurchaseRepository.java
@@ -2,5 +2,5 @@ package com.express.inventory.api.purchases;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseRepository extends JpaRepository<PurchaseEntity, Long> {
+public interface PurchaseRepository extends JpaRepository<PurchaseEntity, Integer> {
 }
